solved bugs

+++ app/src/main/java/com/rescribe/doctor/helpers/doctor_patients/PatientList.java
@@ -31,7 +31,7 @@ public class PatientList implements Parcelable , Comparable<PatientList>
     private String gender;
     @SerializedName("outstandingAmount")
     @Expose
-    private Integer outStandingAmount = 0;
+    private Double outStandingAmount = 0.0;
     @SerializedName("patientId")
     @Expose
     private Integer patientId;
@@ -80,7 +80,7 @@ public class PatientList implements Parcelable , Comparable<PatientList>
         this.age = ((String) in.readValue((String.class.getClassLoader())));
         this.dateOfBirth = ((String) in.readValue((String.class.getClassLoader())));
         this.gender = ((String) in.readValue((String.class.getClassLoader())));
-        this.outStandingAmount = ((Integer) in.readValue((Integer.class.getClassLoader())));
+        this.outStandingAmount = ((Double) in.readValue((Integer.class.getClassLoader())));
         this.patientId = ((Integer) in.readValue((Integer.class.getClassLoader())));
         this.patientPhone = ((String) in.readValue((String.class.getClassLoader())));
         this.patientImageUrl = ((String) in.readValue((String.class.getClassLoader())));
@@ -133,11 +133,11 @@ public class PatientList implements Parcelable , Comparable<PatientList>
         this.gender = gender;
     }
 
-    public Integer getOutStandingAmount() {
+    public Double getOutStandingAmount() {
         return outStandingAmount;
     }
 
-    public void setOutStandingAmount(Integer outStandingAmount) {
+    public void setOutStandingAmount(Double outStandingAmount) {
         this.outStandingAmount = outStandingAmount;
     }
