+++ app/src/main/java/com/rescribe/doctor/model/my_appointments/PatientList.java
@@ -29,7 +29,7 @@ public class PatientList implements Parcelable ,Comparable<PatientList>
     private String gender;
     @SerializedName("outstandingAmount")
     @Expose
-    private int outStandingAmount;
+    private Double outStandingAmount;
     @SerializedName("patientId")
     @Expose
     private Integer patientId;
@@ -95,7 +95,7 @@ public class PatientList implements Parcelable ,Comparable<PatientList>
         this.age = ((String) in.readValue((String.class.getClassLoader())));
         this.dateOfBirth = ((String) in.readValue((String.class.getClassLoader())));
         this.gender = ((String) in.readValue((String.class.getClassLoader())));
-        this.outStandingAmount = ((int) in.readValue((Integer.class.getClassLoader())));
+        this.outStandingAmount = ((Double) in.readValue((Integer.class.getClassLoader())));
         this.patientId = ((Integer) in.readValue((Integer.class.getClassLoader())));
         this.aptId = ((Integer) in.readValue((Integer.class.getClassLoader())));
         this.opdId = ((Integer) in.readValue((Integer.class.getClassLoader())));
@@ -151,11 +151,11 @@ public class PatientList implements Parcelable ,Comparable<PatientList>
         this.gender = gender;
     }
 
-    public int getOutStandingAmount() {
+    public Double getOutStandingAmount() {
         return outStandingAmount;
     }
 
-    public void setOutStandingAmount(int outStandingAmount) {
+    public void setOutStandingAmount(Double outStandingAmount) {
         this.outStandingAmount = outStandingAmount;
     }
